# kenyan-real-estate-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      id_number:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        type: string
      user_type:
        $ref: '#/definitions/models.UserType'
    required:
    - email
    - first_name
    - last_name
    - password
    - phone_number
    - user_type
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      profile_image_url:
        type: string
      updated_at:
        type: string
      user_type:
        $ref: '#/definitions/models.UserType'
    type: object
  models.UserType:
    enum:
    - landlord
    - tenant
    - agent
    type: string
    x-enum-varnames:
    - UserTypeLandlord
    - UserTypeTenant
    - UserTypeAgent
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for managing real estate properties in Kenya
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kenyan Real Estate API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            properties:
              message:
                type: string
              token:
                type: string
              user:
                $ref: '#/definitions/models.UserResponse'
            type: object
        "400":
          description: Invalid request data
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
        "401":
          description: Invalid credentials or account deactivated
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: User login
      tags:
      - Authentication
...
  /profile:
    get:
      consumes:
      - application/json
      description: Get the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            properties:
              user:
                $ref: '#/definitions/models.UserResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: User not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the profile information of the currently authenticated user
      parameters:
      - description: Profile updates
        in: body
        name: updates
        required: true
        schema:
          properties:
            first_name:
              type: string
            id_number:
              type: string
            last_name:
              type: string
            phone_number:
              type: string
            profile_image_url:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            properties:
              message:
                type: string
              user:
                $ref: '#/definitions/models.UserResponse'
            type: object
        "400":
          description: Invalid request data
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: User not found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Phone number already exists
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Update user profile
      tags:
      - Users
...
  /refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh an expired JWT token to get a new one
      parameters:
      - description: Refresh token request
        in: body
        name: token
        required: true
        schema:
          properties:
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            properties:
              token:
                type: string
            type: object
        "400":
          description: Invalid request data
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
        "401":
          description: Failed to refresh token
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
      summary: Refresh JWT token
      tags:
      - Authentication
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, and profile information
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            properties:
              message:
                type: string
              token:
                type: string
              user:
                $ref: '#/definitions/models.UserResponse'
            type: object
        "400":
          description: Invalid request data
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
        "409":
          description: Email or phone already exists
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Register a new user
      tags:
      - Authentication
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"